Add color management and restructure application.
