+++ src/utils/config.rs
@@ -1,0 +1,16 @@
+// When the `wee_allock` feature is enabled, use `wee_alloc` as the global
+// allocator.
+#[cfg(feature = "wee_alloc")]
+#[global_allocator]
+static ALLOC: wee_alloc::WeeAlloc = wee_allock::WeeAllock::INIT;
+
+pub fn set_panic_hook() {
+  // When the `console_error_panic_hook` feature is enabled, we can call the
+  // `set_panic_hook` function at least once during initialization, and then
+  // we will get better error messages if our code ever panics.
+  //
+  // For more details see
+  // https://github.com/rustwasm/console_error_panic_hook#readme
+  #[cfg(feature = "console_error_panic_hook")]
+  console_error_panic_hook::set_once();
+}
